requirements.txt
bidict==0.22.0
click==8.1.3
datetime2==0.8.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
flask-liveserver==0.0.4
Flask-Moment==1.0.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.2.0
Flask-WTF==1.0.1
itsdangerous==2.1.2
mongoengine==0.24.1
pymongo==4.1.1
python-engineio==4.3.2
python-socketio==5.6.0
watchdog==2.1.9
Werkzeug==2.1.2
WTForms==3.0.1
